+++ src/Colibri/Configuration/Models/ClusterCfg.cs
@@ -34,7 +34,8 @@ public sealed class ClusterCfg
             .ToArray();
     } = null!;
     
-    public LoadBalancingCfg? LoadBalancing { get; set; }
+    public LoadBalancingCfg? LoadBalancer { get; set; }
     public CircuitBreakerCfg? CircuitBreaker { get; set; }
     public RetryCfg? Retry { get; set; }
+    public AuthCfg? Authorization { get; set; }
 }