+++ tests/Unit/Tests/Validators/ClusterValidatorTests.cs
@@ -302,4 +302,57 @@ public sealed class ClusterValidatorTests
         // Assert
         Assert.True(result);
     }
+
+    [Theory]
+    [InlineData("")]
+    [InlineData(" ")]
+    public void AuthAlgorithmIsNotEmpty_WhenAuthAlgorithmIsEmpty_ShouldReturnFalse(string algorithm)
+    {
+        // Act
+        var result = _validator.AuthAlgorithmIsNotEmpty(algorithm);
+        
+        // Assert
+        Assert.False(result);
+    }
+
+    [Theory]
+    [InlineData("rs256")]
+    [InlineData("hs256")]
+    [InlineData("es256")]
+    [InlineData("BELIBERDA")]
+    public void AuthAlgorithmIsNotEmpty_WhenAuthAlgorithmIsNotEmpty_ShouldReturnTrue(string algorithm)
+    {
+        // Act
+        var result = _validator.AuthAlgorithmIsNotEmpty(algorithm);
+        
+        // Assert
+        Assert.True(result);
+    }
+
+    [Theory]
+    [InlineData("rs255")]
+    [InlineData("ES221")]
+    [InlineData("hs512")]
+    [InlineData("BELIBERDA")]
+    public void AuthAlgorithmIsValid_WhenAuthAlgorithmIsInvalid_ShouldReturnFalse(string algorithm)
+    {
+        // Act
+        var result = _validator.AuthAlgorithmIsValid(algorithm);
+        
+        // Assert
+        Assert.False(result);
+    }
+    
+    [Theory]
+    [InlineData("rs256")]
+    [InlineData("hs256")]
+    [InlineData("es256")]
+    public void AuthAlgorithmIsValid_WhenAuthAlgorithmIsValid_ShouldReturnTrue(string algorithm)
+    {
+        // Act
+        var result = _validator.AuthAlgorithmIsValid(algorithm);
+        
+        // Assert
+        Assert.True(result);
+    }
 }